OADP-659 Move StripDefaultPorts to pkg/common.
