+++ pkg/bucket/aws.go
@@ -2,12 +2,9 @@ package bucket
 
 import (
 	"fmt"
-	"net/http"
-	"net/url"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
-	"github.com/aws/aws-sdk-go/aws/request"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
@@ -172,19 +169,4 @@ func (a awsBucketClient) Delete() (bool, error) {
 	}
 
 	return true, nil
-}
-
-// StripDefaultPorts removes port 80 from HTTP URLs and 443 from HTTPS URLs.
-// Defer to the actual AWS SDK implementation to match its behavior exactly.
-func StripDefaultPorts(fromUrl string) (string, error) {
-	u, err := url.Parse(fromUrl)
-	if err != nil {
-		return "", err
-	}
-	r := http.Request{
-		URL: u,
-	}
-	request.SanitizeHostForHeader(&r)
-	r.URL.Host = r.Host
-	return r.URL.String(), nil
 }
